코드 정리

+++ CSphere.h
@@ -3,8 +3,8 @@
 
 
 #include "d3dUtility.h"
-#define M_RADIUS 0.21   // ball radius (= 2.85cm, 1cm = 0.073684)
-#define DECREASE_RATE 0.9982 //rate of speed decrease
+#define M_RADIUS 0.21f   // ball radius (= 2.85cm, 1cm = 0.073684)
+#define DECREASE_RATE 0.9982f //rate of speed decrease
 
 class CSphere {
 private:
@@ -27,13 +27,13 @@ public:
 
 	void ballUpdate(float timeDiff);
 
-	double getVelocity_X() const { return this->m_velocity_x; }
-	double getVelocity_Y() const { return this->m_velocity_y; }
-	double getVelocity_Z() const { return this->m_velocity_z; }
+	float getVelocity_X() const { return this->m_velocity_x; }
+	float getVelocity_Y() const { return this->m_velocity_y; }
+	float getVelocity_Z() const { return this->m_velocity_z; }
 	D3DXVECTOR3 getVelocity() const { return D3DXVECTOR3(m_velocity_x, m_velocity_y, m_velocity_z); }
 
-	inline void setPower(double vx, double vz) { setPower(vx, getVelocity_Y(), vz); }
-	inline void setPower(double vx, double vy, double vz);
+	inline void setPower(float vx, float vz) { setPower(vx, getVelocity_Y(), vz); }
+	inline void setPower(float vx, float vy, float vz);
 	void setCenter(float x, float y, float z);
 
 	float getRadius(void)  const { return this->m_radius; }
